move out engine utils to own utils file as step 1 of making the engine smaller
