+++ lib/core/engine.js
@@ -21,14 +21,12 @@ const preURL = require('../support/preURL');
 const extensionSetup = require('../support/extensionSetup');
 const finetuneVideo = require('../support/video/scripts/finetuneVideo');
 const connectivity = require('../support/connectivity');
-const path = require('path');
 const util = require('../support/util');
 const filterWhitelisted = require('../support/userTiming').filterWhitelisted;
-const moment = require('moment');
 const get = require('lodash.get');
 const xvfb = require('../support/video/xvfb');
 const extensionServer = require('../support/extensionServer');
-const forEach = require('lodash.foreach');
+const engineUtils = require('../support/engineUtils');
 
 const defaults = {
   scripts: [],
@@ -36,107 +34,18 @@ const defaults = {
   delay: 0
 };
 
-function loadPrePostScripts(scripts) {
-  return util.toArray(scripts).map(script => {
-    try {
-      return require(path.resolve(script));
-    } catch (e) {
-      throw new Error(
-        "Couldn't run pre/post script file: " + path.resolve(script) + ' ' + e
-      );
-    }
-  });
-}
-
-function jsonifyVisualProgress(visualProgress) {
-  // Original data looks like
-  //  "0=0%, 1500=81%, 1516=81%, 1533=84%, 1550=84%, 1566=84%, 1600=95%, 1683=95%, 1833=100%"
-  const progress = visualProgress.split(',');
-  const visualProgressJSON = {};
-  forEach(progress, value => {
-    const eachMetric = value.split('=');
-    visualProgressJSON[eachMetric[0].replace(' ', '')] = Number(
-      eachMetric[1].replace('%', '')
-    );
-  });
-  return visualProgressJSON;
-}
-
-function addExtrasToHAR(harPage, visualMetricsData, timings, options) {
-  const _meta = (harPage._meta = {});
-  _meta.connectivity = get(options, 'connectivity.profile', 'native');
-  const harPageTimings = harPage.pageTimings;
-
-  const _visualMetrics = (harPage._visualMetrics = {});
-
-  // We add the timings both as a hidden field and add
-  // in pageTimings so we can easily show them in PerfCascade
-  if (visualMetricsData) {
-    harPageTimings._firstVisualChange = visualMetricsData.FirstVisualChange;
-    harPageTimings._lastVisualChange = visualMetricsData.LastVisualChange;
-    harPageTimings._visualComplete85 = visualMetricsData.VisualComplete85;
-    harPageTimings._visualComplete95 = visualMetricsData.VisualComplete95;
-    harPageTimings._visualComplete99 = visualMetricsData.VisualComplete99;
-    _visualMetrics.FirstVisualChange = visualMetricsData.FirstVisualChange;
-    _visualMetrics.SpeedIndex = visualMetricsData.SpeedIndex;
-    _visualMetrics.VisualComplete85 = visualMetricsData.VisualComplete85;
-    _visualMetrics.VisualComplete95 = visualMetricsData.VisualComplete95;
-    _visualMetrics.VisualComplete99 = visualMetricsData.VisualComplete99;
-    _visualMetrics.LastVisualChange = visualMetricsData.LastVisualChange;
-
-    // Make the visual progress structure more JSON
-    _visualMetrics.VisualProgress = jsonifyVisualProgress(
-      visualMetricsData.VisualProgress
-    );
-  } else if (timings && timings.firstPaint) {
-    // only add first paint if we don't have visual metrics
-    harPageTimings._firstPaint = timings.firstPaint;
-  }
-  if (timings && timings.pageTimings) {
-    harPageTimings._domInteractiveTime = timings.pageTimings.domInteractiveTime;
-    harPageTimings._domContentLoadedTime =
-      timings.pageTimings.domContentLoadedTime;
-  }
-}
-
-function calculateViewport(options) {
-  const emulatedWidth = get(options, 'chrome.mobileEmulation.width');
-  const emulatedHeight = get(options, 'chrome.mobileEmulation.height');
-  // you cannot set the width/height for phone so just keep the viewport undefined
-  if (get(options, 'chrome.android.package')) {
-    return;
-  }
-
-  if (emulatedWidth && emulatedHeight) {
-    return `${emulatedWidth}x${emulatedHeight}`;
-  }
-
-  if (
-    /^\d+x\d+$/.test(options.viewPort) ||
-    (options.viewPort === 'maximize' && !options.xvfb)
-  ) {
-    return options.viewPort;
-  }
-
-  return '1200x960';
-}
-
-function timestamp() {
-  return moment().format();
-}
-
 class Engine {
   constructor(options) {
     try {
-      options.preScript = loadPrePostScripts(options.preScript);
-      options.postScript = loadPrePostScripts(options.postScript);
+      options.preScript = engineUtils.loadPrePostScripts(options.preScript);
+      options.postScript = engineUtils.loadPrePostScripts(options.postScript);
     } catch (e) {
       log.error(e.message);
       throw e;
     }
 
     this.options = merge({}, defaults, options);
-    this.options.viewPort = calculateViewport(this.options);
+    this.options.viewPort = engineUtils.calculateViewport(this.options);
     this.runDelegate = engineDelegate.createDelegate(this.options);
   }
 
@@ -251,7 +160,7 @@ class Engine {
           version
         },
         url,
-        timestamp: timestamp(),
+        timestamp: engineUtils.timestamp(),
         connectivity: {
           engine: get(options, 'connectivity.engine'),
           profile: get(options, 'connectivity.profile')
@@ -328,7 +237,7 @@ class Engine {
         )
         .tap(() => runDelegate.onStartIteration(runner, index))
         .tap(result => {
-          result.timestamp = timestamp();
+          result.timestamp = engineUtils.timestamp();
         })
         .tap(() => runner.loadAndWait(url, options.pageCompleteCheck))
         .tap(result => {
@@ -474,7 +383,12 @@ class Engine {
             const page = result.har.log.pages[run];
             const visualMetric = result.visualMetrics[run];
             const browserScript = result.browserScripts[run] || {};
-            addExtrasToHAR(page, visualMetric, browserScript.timings, options);
+            engineUtils.addExtrasToHAR(
+              page,
+              visualMetric,
+              browserScript.timings,
+              options
+            );
           }
         }
       })
